(refactor requests): contract requests
