+++ src/domain/legacy.js
@@ -1,6 +1,6 @@
 import { ethers } from "ethers";
 import { gql } from "@apollo/client";
-import { useState, useEffect, useMemo, useCallback, useRef } from "react";
+import { useCallback, useEffect, useMemo, useRef, useState } from "react";
 import { Token as UniToken } from "@uniswap/sdk-core";
 import { Pool } from "@uniswap/v3-sdk";
 import useSWR from "swr";
@@ -18,36 +18,33 @@ import PositionRouter from "../abis/PositionRouter.json";
 import { getContract } from "../config/Addresses";
 import { getConstant } from "../config/chains";
 import {
-  UI_VERSION,
   ARBITRUM,
+  ARBITRUM_TESTNET,
   AVALANCHE,
-  // DEFAULT_GAS_LIMIT,
   bigNumberify,
-  getExplorerUrl,
+  DECREASE,
+  expandDecimals,
+  getHighExecutionFee,
+  getInfoTokens,
+  getOrderKey,
+  getProvider,
   getServerBaseUrl,
   getServerUrl,
-  setGasPrice,
-  getGasLimit,
-  replaceNativeTokenAddress,
-  getProvider,
-  getOrderKey,
-  fetcher,
-  parseValue,
-  expandDecimals,
-  getInfoTokens,
-  helperToast,
   getUsd,
-  USD_DECIMALS,
-  getHighExecutionFee,
-  SWAP,
   INCREASE,
-  DECREASE,
-  ARBITRUM_TESTNET,
+  parseValue,
+  replaceNativeTokenAddress,
+  SWAP,
+  UI_VERSION,
+  USD_DECIMALS,
 } from "../lib/legacy";
-import { getTokens, getTokenBySymbol, getWhitelistedTokens } from "../config/Tokens";
+import { getTokenBySymbol, getTokens, getWhitelistedTokens } from "../config/Tokens";
 
-import { nissohGraphClient, arbitrumGraphClient, avalancheGraphClient } from "./common";
+import { arbitrumGraphClient, avalancheGraphClient, nissohGraphClient } from "./common";
 import { groupBy } from "lodash";
+import { fetcher } from "../lib/contracts/fetcher";
+import { callContract } from "../lib/contracts/callContract";
+
 export * from "./prices";
 
 const { AddressZero } = ethers.constants;
@@ -975,166 +972,4 @@ export function executeIncreaseOrder(chainId, library, account, index, feeReceiv
 
 export function executeDecreaseOrder(chainId, library, account, index, feeReceiver, opts) {
   return _executeOrder(chainId, library, "executeDecreaseOrder", account, index, feeReceiver, opts);
-}
-
-const NOT_ENOUGH_FUNDS = "NOT_ENOUGH_FUNDS";
-const USER_DENIED = "USER_DENIED";
-const SLIPPAGE = "SLIPPAGE";
-const RPC_ERROR = "RPC_ERROR";
-
-const TX_ERROR_PATTERNS = {
-  [NOT_ENOUGH_FUNDS]: [
-    { msg: "not enough funds for gas" },
-    { msg: "failed to execute call with revert code InsufficientGasFunds" },
-  ],
-  [USER_DENIED]: [{ msg: "User denied transaction signature" }],
-  [SLIPPAGE]: [{ msg: "Router: mark price lower than limit" }, { msg: "Router: mark price higher than limit" }],
-  [RPC_ERROR]: [
-    // @see https://eips.ethereum.org/EIPS/eip-1474#error-codes
-    { code: -32005 },
-    { msg: "Non-200 status code" },
-    { msg: "Request limit exceeded" },
-    { msg: "Internal JSON-RPC error" },
-    { msg: "Response has no error or result" },
-    { msg: "couldn't connect to the network" },
-  ],
-};
-
-export function extractError(ex) {
-  if (!ex) {
-    return [];
-  }
-
-  const message = ex.data?.message || ex.message;
-  const code = ex.code;
-
-  if (!message && !code) {
-    return [];
-  }
-
-  for (const [type, patterns] of Object.entries(TX_ERROR_PATTERNS)) {
-    for (const pattern of patterns) {
-      const matchCode = pattern.code && code === pattern.code;
-      const matchMessage = pattern.msg && message && message.includes(pattern.msg);
-
-      if (matchCode || matchMessage) {
-        return [message, type, ex.data];
-      }
-    }
-  }
-
-  return [message, null, ex.data];
-}
-
-function ToastifyDebug(props) {
-  const [open, setOpen] = useState(false);
-  return (
-    <div className="Toastify-debug">
-      <span className="Toastify-debug-button" onClick={() => setOpen((old) => !old)}>
-        {open ? "Hide error" : "Show error"}
-      </span>
-      {open && <div className="Toastify-debug-content">{props.children}</div>}
-    </div>
-  );
-}
-
-export async function callContract(chainId, contract, method, params, opts) {
-  try {
-    if (!Array.isArray(params) && typeof params === "object" && opts === undefined) {
-      opts = params;
-      params = [];
-    }
-    if (!opts) {
-      opts = {};
-    }
-
-    const txnOpts = {};
-
-    if (opts.value) {
-      txnOpts.value = opts.value;
-    }
-
-    txnOpts.gasLimit = opts.gasLimit ? opts.gasLimit : await getGasLimit(contract, method, params, opts.value);
-
-    await setGasPrice(txnOpts, contract.provider, chainId);
-
-    const res = await contract[method](...params, txnOpts);
-    const txUrl = getExplorerUrl(chainId) + "tx/" + res.hash;
-    const sentMsg = opts.sentMsg || "Transaction sent.";
-    helperToast.success(
-      <div>
-        {sentMsg}{" "}
-        <a href={txUrl} target="_blank" rel="noopener noreferrer">
-          View status.
-        </a>
-        <br />
-      </div>
-    );
-    if (opts.setPendingTxns) {
-      const message = opts.hideSuccessMsg ? undefined : opts.successMsg || "Transaction completed!";
-      const pendingTxn = {
-        hash: res.hash,
-        message,
-      };
-      opts.setPendingTxns((pendingTxns) => [...pendingTxns, pendingTxn]);
-    }
-    return res;
-  } catch (e) {
-    let failMsg;
-    let autoCloseToast = 5000;
-
-    const [message, type, errorData] = extractError(e);
-    switch (type) {
-      case NOT_ENOUGH_FUNDS:
-        failMsg = (
-          <div>
-            There is not enough ETH in your account on Arbitrum to send this transaction.
-            <br />
-            <br />
-            <a href={"https://arbitrum.io/bridge-tutorial/"} target="_blank" rel="noopener noreferrer">
-              Bridge ETH to Arbitrum
-            </a>
-          </div>
-        );
-        break;
-      case USER_DENIED:
-        failMsg = "Transaction was cancelled.";
-        break;
-      case SLIPPAGE:
-        failMsg =
-          'The mark price has changed, consider increasing your Allowed Slippage by clicking on the "..." icon next to your address.';
-        break;
-      case RPC_ERROR:
-        autoCloseToast = false;
-
-        const originalError = errorData?.error?.message || errorData?.message || message;
-
-        failMsg = (
-          <div>
-            Transaction failed due to RPC error.
-            <br />
-            <br />
-            Please try changing the RPC url in your wallet settings.{" "}
-            <a href="https://gmxio.gitbook.io/gmx/trading#backup-rpc-urls" target="_blank" rel="noopener noreferrer">
-              More info
-            </a>
-            <br />
-            {originalError && <ToastifyDebug>{originalError}</ToastifyDebug>}
-          </div>
-        );
-        break;
-      default:
-        autoCloseToast = false;
-
-        failMsg = (
-          <div>
-            {opts.failMsg || "Transaction failed"}
-            <br />
-            {message && <ToastifyDebug>{message}</ToastifyDebug>}
-          </div>
-        );
-    }
-    helperToast.error(failMsg, { autoClose: autoCloseToast });
-    throw e;
-  }
 }
