+++ src/pages/Exchange/Exchange.js
@@ -18,7 +18,6 @@ import {
   formatAmount,
   bigNumberify,
   getTokenInfo,
-  fetcher,
   getPositionKey,
   getPositionContractKey,
   getLeverage,
@@ -52,6 +51,7 @@ import Tab from "../../components/Tab/Tab";
 import Footer from "../../components/Footer/Footer";
 
 import "./Exchange.css";
+import { fetcher } from "../../lib/contracts/fetcher";
 const { AddressZero } = ethers.constants;
 
 const PENDING_POSITION_VALID_DURATION = 600 * 1000;
